requirements.txt
uvicorn==0.40.0
fastapi==0.128.0
python-dotenv==1.2.1
langchain-core==1.2.9
langchain-openai==1.1.8
langgraph==1.0.8
pinecone==7.3.0
langchain-pinecone==0.2.13